# Conflicts:
#	src/app/app-routing.module.ts
#	src/app/contacts/contact-detail/contact-detail.component.ts

+++ src/app/contacts/contact-list/contact-list.component.ts
@@ -8,39 +8,22 @@ import {ContactService} from '../contact.service';
   styleUrls: ['./contact-list.component.css']
 })
 export class ContactListComponent implements OnInit {
-  contacts: Contact[];
-  filteredContacts: Contact[];
+  public contacts: Contact[];
+  public contact: Contact;
+  public searchTerm: string;
 
-  private contact: Contact;
-  private searchTerm: string;
-
-  getSearchTerm() {
-    return this.searchTerm;
-  }
-
-  setSearchTerm(value: string) {
-    this.searchTerm = value;
-    this.filteredContacts = this.filterContacts(value);
-  }
 
   constructor(private contactService: ContactService) {
   }
 
   ngOnInit() {
     this.contactService.getReq()
-      .subscribe((contact: Contact[]) => this.contacts = contact);
-    this.contactService.getReq().subscribe(result => console.log(result));
+      .subscribe((contact: Contact[]) => this.contacts = contact.sort((a, b) => (
+        a['surname'] > b['surname']) ? 1 : ((b['surname'] > a['surname']) ? -1 : 0)));
 
-    this.contactService.deleteContact(this.contact.contactID).subscribe();
-    this.contactService.addContact(this.contact).subscribe(contact => this.contacts.push(contact));
-    this.filteredContacts = this.contacts;
-
+    if (this.contact) {
+      this.contactService.deleteContact(this.contact.contactID).subscribe();
+      this.contactService.addContact(this.contact).subscribe(contact => this.contacts.push(contact));
+    }
   }
-
-  private filterContacts(searchString: string) {
-    return this.contacts.filter(contact =>
-      contact.firstName.toLowerCase().indexOf(searchString));
-  }
-
-
 }
